requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
firebase-admin==6.0.1
opencv-python-headless==4.8.0.74
face-recognition==1.3.0
mysql-connector-python==8.1.0
numpy==1.24.3
